Added 'Synthetic' Samples
The main change for this was adding the 'SMOTE' algorithm to the classifier.
One weakness with the dataset is that most tweets have the 'others' emotion. This means that a classifier can appear superficially accurate merely by labelling all tweets as 'others'. 
For this step, the performance of the algorithm on under-represented classes has been greatly improved without hurting overall accuracy.